requirements.txt
selenium==4.18.1
pandas==2.2.1
SQLAlchemy==2.0.27
pymysql==1.1.0
python-dotenv==1.0.1
openpyxl==3.1.2
ipykernel==6.29.5
